AgentSystemV2/requirements.txt
# AgentSystemV2 Requirements

# Core dependencies
agno>=1.0.0
python-dotenv>=1.0.0

# Async support
asyncio-signal>=0.1.0

# Data handling
dataclasses-json>=0.6.0

# Logging and monitoring
python-json-logger>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Optional: Enhanced features
rich>=13.0.0  # For better CLI output
typer>=0.9.0  # For CLI interface
pydantic>=2.0.0  # For enhanced data validation